fix: 修复 host 侧单元测试的编译错误和断言失效问题
本次提交修复了多个 host 侧单元测试的问题，包括：
1. 修正了整数运算溢出、类型不匹配等编译错误
2. 删除了依赖硬件 I/O 的测试用例，等价断言迁移至 kernel_test 载体
3. 修复了断言与实现语义不符的问题，如地址范围、字段取值等
4. 移除了无意义的恒真断言
5. 修正了系统调用参数校验逻辑，符合 POSIX 规范
6. 调整了 aarch64 构建目标为 softfloat，适配内核零浮点策略
7. 修复了 eBPF 验证器的指令初始化检查逻辑，避免误拒合法指令
8. 修正了错误码映射，将 NotSupported 映射为 ENOTSUP 而非 ENOSYS
9. 修复了 IPC 管道的文件描述符校验顺序，避免掩盖无效 fd 错误
10. 修正了 TSS 结构的字段取值和对齐断言，符合硬件规范
11. 修复了哈希链验证逻辑，避免覆盖有效节点导致校验失败

+++ src/kernel/framework/mm/copy_user.rs
@@ -662,8 +662,9 @@ mod tests {
         assert!(!is_user_ptr(USER_ADDR_MAX));
         assert!(!is_user_ptr(USER_ADDR_MAX + 1));
 
+        // USER_ADDR_MAX 为 u64 地址, 与 usize 型常量同转 u64 后相减.
         assert!(is_user_buf(
-            USER_ADDR_MAX - PAGE_SIZE as usize,
+            USER_ADDR_MAX - PAGE_SIZE as u64,
             PAGE_SIZE as usize
         ));
         assert!(!is_user_buf(USER_ADDR_MAX - 100, 200));
